refactor: rename `core` package to `cuore` to align with Rust workspace
