+++ src/mnemolet/cli/commands/list.py
@@ -3,7 +3,7 @@ import click
 from mnemolet.config import (
     QDRANT_URL,
 )
-from mnemolet.core.utils.qdrant import QdrantManager
+from mnemolet.cuore.utils.qdrant import QdrantManager
 
 from .utils import requires_qdrant
