+++ src/mnemolet/cli/main.py
@@ -8,7 +8,7 @@ import click
 from mnemolet.config import (
     QDRANT_URL,
 )
-from mnemolet.core.utils.qdrant import QdrantManager
+from mnemolet.cuore.utils.qdrant import QdrantManager
 
 logger = logging.getLogger(__name__)
