+++ src/mnemolet/cuore/embeddings/local_llm_embed.py
