+++ src/mnemolet/cuore/health/ollama.py
