+++ src/mnemolet/cuore/ingestion/extractors/base.py
