+++ src/mnemolet/cuore/query/generation/local_generator.py
