+++ src/mnemolet/cuore/storage/base.py
